fix(pencil): Fix repeated headers to be bounded properly
Previously, the field lookup occurred from the beginning of the line
so would repeated find the same line location if the same string or a substring
of a header was searched more than once.

Instead track the current end index of the previous field to base the search
state location on.